Add 'env' property support to configuration.
Variables specified under the `env` property in the root of the yaml
configuration file will now be set in the pups environment during the
initialization process. Any existing environment variables are
overwritten by matching variables specified in the config `env` property.

The integration test was updated to exercise the params and env
codepath as this covers the new functionality.